Broke out modules getSpannedTitle and replaceTitleWithSpanned from main.js #28
